% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfrm.R
\name{bfrm}
\alias{bfrm}
\title{Fit Bayesian Linear Mixed Model with JZS Prior}
\usage{
bfrm(formula, data, family = gaussian(), prior_structure = "jzs",
  prior_arg = list(r_fixed = 0.5, r_random = 0.5), ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}

\item{data}{An object of class \code{data.frame} containing all observations and
variables used in the model.}

\item{family}{Currently only \code{\link[=gaussian]{gaussian()}} is supported.}

\item{prior_structure}{character string. Currently only \code{'jzs'} (for
Jeffreys-Zellner-Siow) is supported.}

\item{prior_arg}{\code{list} of additional arguments specific to
\code{prior_structure}. For \code{'jzs'}, these can be \code{r_fixed} and \code{r_random}.}

\item{...}{Further arguments passed to \code{\link[brms:brm]{brms::brm()}} such as \code{iter},
\code{warmup}, or \code{cores}.}
}
\description{
Fit Bayesian Linear Mixed Model with JZS Prior
}
