% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_like_functions.R
\name{make_stancode_bfrms}
\alias{make_stancode_bfrms}
\alias{brms_like}
\alias{make_standata_bfrms}
\title{brms-like Functions}
\usage{
make_stancode_bfrms(formula, data, family = gaussian(),
  prior_structure = "jzs", prior_arg = list(r_fixed = 0.5, r_random =
  1), ...)

make_standata_bfrms(formula, data, family = gaussian(),
  prior_structure = "jzs", prior_arg = list(r_fixed = 0.5, r_random =
  1), ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}

\item{data}{An object of class \code{data.frame} containing all observations and
variables used in the model.}

\item{family}{Currently only \code{\link[=gaussian]{gaussian()}} is supported.}

\item{prior_structure}{character string. Currently only \code{'jzs'} (for
Jeffreys-Zellner-Siow) is supported.}

\item{prior_arg}{\code{list} of additional arguments specific to
\code{prior_structure}. For \code{'jzs'}, these can be \code{r_fixed} and \code{r_random}.}

\item{...}{further arguments passed to the corresponding \code{brms} function. See
their help pages for details.}
}
\value{
\code{make_stancode_bfrms} returns the same as \code{\link[brms:make_stancode]{brms::make_stancode()}} and
\code{make_standata_bfrms} returns the same as \code{\link[brms:make_standata]{brms::make_standata()}}.
}
\description{
Generate code or data for \code{brms} models using JSZ-priors to be
passed to \code{Stan}. \code{make_stancode_bfrms} is the \code{bfrms} alias of
\code{\link[brms:make_stancode]{brms::make_stancode()}} and \code{make_standata_bfrms} is the \code{bfrms} alias of
\code{\link[brms:make_standata]{brms::make_standata()}}.
}
