% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.bayes.R
\name{contr.bayes}
\alias{contr.bayes}
\title{Orthonormal Contrast Matrices for Bayesian Estimation}
\usage{
contr.bayes(n, contrasts = TRUE)
}
\arguments{
\item{n}{a vector of levels for a factor, or the number of levels.}

\item{contrasts}{logical indicating whether contrasts should be computed.}
}
\value{
A \code{matrix} with n rows and k columns, with k=n-1 if contrasts is
\code{TRUE`` and k=n if contrasts is }FALSE`.
}
\description{
Returns a design or model matrix of orthonormal contrasts such that the
marginal prior on all effects is identical. Implementation follows the
description in Rouder, Morey, Speckman, & Province (2012, p. 363).
}
\examples{



contr.bayes(2)  ## Q_2 in Rouder et al. (2012, p. 363)
#            [,1]
# [1,] -0.7071068
# [2,]  0.7071068

contr.bayes(5)  ## equivalent to Q_5 in Rouder et al. (2012, p. 363)

## check decomposition
Q3 <- contr.bayes(3)
Q3 \%*\% diag(2) \%*\% t(Q3)
## 2/3 on diagonal and -1/3 on off-diagonal elements
}
\references{
Rouder, J. N., Morey, R. D., Speckman, P. L., & Province, J. M.
(2012). Default Bayes factors for ANOVA designs. \emph{Journal of Mathematical
Psychology}, 56(5), 356-374. https://doi.org/10.1016/j.jmp.2012.08.001
}
